# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual Keyboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RecordingEvent {
  key: string;
  timestamp: number;
  instrument: string;
  velocity: number;
  rgbColor: string;
  glowIntensity: number;
}

export interface Recording {
  id: number;
  name: string;
  sessionId: string;
  events: RecordingEvent[];
  duration: number;
  bpm: number;
  createdAt: string;
  updatedAt: string;
}

export interface RecordingInput {
  name: string;
  sessionId: string;
  events: RecordingEvent[];
  duration: number;
  bpm: number;
}

export type PresetSettings = { [key: string]: unknown };

export interface Preset {
  id: number;
  name: string;
  sessionId: string;
  settings: PresetSettings;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export type PresetInputSettings = { [key: string]: unknown };

export interface PresetInput {
  name: string;
  sessionId: string;
  settings: PresetInputSettings;
  isPublic?: boolean;
}

export type ListRecordingsParams = {
sessionId?: string;
};

export type ListPresetsParams = {
sessionId?: string;
};