# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Virtual Keyboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recordings
    description: Recording operations
  - name: presets
    description: Preset operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /recordings:
    get:
      operationId: listRecordings
      tags: [recordings]
      summary: List recordings
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
    post:
      operationId: createRecording
      tags: [recordings]
      summary: Save a recording
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordingInput"
      responses:
        "201":
          description: Recording saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"

  /recordings/{id}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get a recording by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recording details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRecording
      tags: [recordings]
      summary: Delete a recording
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /presets:
    get:
      operationId: listPresets
      tags: [presets]
      summary: List presets
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Preset"
    post:
      operationId: createPreset
      tags: [presets]
      summary: Save a preset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresetInput"
      responses:
        "201":
          description: Preset saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preset"

  /presets/{id}:
    delete:
      operationId: deletePreset
      tags: [presets]
      summary: Delete a preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    RecordingEvent:
      type: object
      properties:
        key:
          type: string
        timestamp:
          type: number
        instrument:
          type: string
        velocity:
          type: number
        rgbColor:
          type: string
        glowIntensity:
          type: number
      required:
        - key
        - timestamp
        - instrument
        - velocity
        - rgbColor
        - glowIntensity

    Recording:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sessionId:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/RecordingEvent"
        duration:
          type: integer
        bpm:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sessionId
        - events
        - duration
        - bpm
        - createdAt
        - updatedAt

    RecordingInput:
      type: object
      properties:
        name:
          type: string
        sessionId:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/RecordingEvent"
        duration:
          type: integer
        bpm:
          type: integer
      required:
        - name
        - sessionId
        - events
        - duration
        - bpm

    Preset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sessionId:
          type: string
        settings:
          type: object
          additionalProperties: true
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sessionId
        - settings
        - isPublic
        - createdAt
        - updatedAt

    PresetInput:
      type: object
      properties:
        name:
          type: string
        sessionId:
          type: string
        settings:
          type: object
          additionalProperties: true
        isPublic:
          type: boolean
      required:
        - name
        - sessionId
        - settings