# lib/api-zod/src/generated/types/preset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual Keyboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PresetSettings } from './presetSettings';

export interface Preset {
  id: number;
  name: string;
  sessionId: string;
  settings: PresetSettings;
  isPublic: boolean;
  createdAt: Date;
  updatedAt: Date;
}