# lib/api-zod/src/generated/types/presetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual Keyboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PresetInputSettings } from './presetInputSettings';

export interface PresetInput {
  name: string;
  sessionId: string;
  settings: PresetInputSettings;
  isPublic?: boolean;
}