# lib/api-zod/src/generated/types/recording.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual Keyboard API
 * OpenAPI spec version: 0.1.0
 */
import type { RecordingEvent } from './recordingEvent';

export interface Recording {
  id: number;
  name: string;
  sessionId: string;
  events: RecordingEvent[];
  duration: number;
  bpm: number;
  createdAt: Date;
  updatedAt: Date;
}