# lib/api-zod/src/generated/types/recordingEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Virtual Keyboard API
 * OpenAPI spec version: 0.1.0
 */

export interface RecordingEvent {
  key: string;
  timestamp: number;
  instrument: string;
  velocity: number;
  rgbColor: string;
  glowIntensity: number;
}